+++ windlast_UI/static/js/help_content/tor.js
@@ -87,7 +87,7 @@ export const TOR_HELP_PAGES = [
         Gummimatten blau markiert.<br>
         Da die Gummimatte Einfluss auf den Reibwert hat und nicht für alle Materialpaarungen Reibwerte in den zugrunde gelegten Normen
         ([[norm:DIN_EN_13814_2005|DIN EN 13814:2005-06]] und [[norm:DIN_EN_17879_2024|DIN EN 17879:2024-08]]) angegeben sind, können die Auswahlmöglichkeiten in diesem Feld
-        je nach ausgewählter [[tor:bodenplatte|Bodenplatte]] eingeschränkt sein. Die Auswahl der Gummimatte kann auch Einfluss auf
+        je nach ausgewählter [[tor:bodenplatte|Bodenplatte]] eingeschränkt sein oder automatisch geändert werden. Die Auswahl der Gummimatte kann auch Einfluss auf
         auf die Auswahlmöglichkeiten im [[tor:untergrund|Untergrund-Feld]] haben.
       </p>
     `
@@ -103,7 +103,7 @@ export const TOR_HELP_PAGES = [
         Wählt hier den Untergrund aus, auf dem die Konstruktion steht. Der Untergrund beeinflusst den Reibwert zwischen Bodenplatte und Untergrund. Dieser wird anhand der
         [[norm:DIN_EN_13814_2005|DIN EN 13814:2005-06]] bzw. [[norm:DIN_EN_17879_2024|DIN EN 17879:2024-08]] automatisch bestimmt.<br>
         Da der Reibwert nicht für alle Materialpaarungen in den zugrunde gelegten Normen angegeben ist, können die Auswahlmöglichkeiten in diesem Feld
-        je nach ausgewählter [[tor:bodenplatte|Bodenplatte]] und [[tor:gummimatte|Gummimatte]] eingeschränkt sein.
+        je nach ausgewählter [[tor:bodenplatte|Bodenplatte]] und [[tor:gummimatte|Gummimatte]] eingeschränkt sein oder automatisch geändert werden.
       </p>
     `
   },